Newfix03 (fmidue/ba-tongxi-xu#57)
1, add a description for issue 4 to describe associative operators

2,delete not necessary fromIntegral

3, change some names to make it clear

4, add a test for issue 4 and change some tests for issue 2

+++ src/Tasks/SuperfluousBrackets/PrintSuperfluousBrackets.hs
@@ -36,8 +36,8 @@ allocateBracketToSubtree a b (oper, nowSerial) brackets fatherOperator serial
     | Just nowSerial == serial = do
         formula <- allocateBracketToSubtree a b (oper, nowSerial) brackets fatherOperator Nothing
         return ("(" ++ formula ++ ")")
-    | otherwise = let rightNodes = fromIntegral (treeNodes b)
-                      leftNodes = fromIntegral (treeNodes a)
+    | otherwise = let rightNodes =  treeNodes b
+                      leftNodes =  treeNodes a
                 in  do
                 ifUsebrackets <- frequency [(fromIntegral brackets, return True), (fromIntegral (rightNodes + leftNodes + 1 - brackets), return False)]
                 let brackets' = if ifUsebrackets then brackets - 1 else brackets
